move packages around preset/dev
