+++ linen-modern/src/main/scala/x7c1/linen/modern/init/dev/ChannelSelector.scala
@@ -4,7 +4,7 @@ import android.app.AlertDialog
 import android.content.DialogInterface.{OnClickListener, OnMultiChoiceClickListener}
 import android.content.{Context, DialogInterface}
 import x7c1.linen.database.LinenOpenHelper
-import x7c1.linen.modern.accessor.dev.ChannelNameAccessor
+import x7c1.linen.repository.dev.ChannelNameAccessor
 import x7c1.wheat.macros.logger.Log
 
 class ChannelSelector private (
