+++ linen-modern/src/main/scala/x7c1/linen/modern/init/settings/preset/AttachSourceDialog.scala
@@ -13,16 +13,15 @@ import android.widget.{Button, CompoundButton}
 import x7c1.linen.database.{ChannelSourceMapKey, ChannelSourceMapParts, LinenOpenHelper}
 import x7c1.linen.domain.Date
 import x7c1.linen.glue.res.layout.{SettingSourceAttach, SettingSourceAttachRow, SettingSourceAttachRowItem}
-import x7c1.linen.modern.accessor.setting.{ChannelToAttach, ChannelsToAttachAccessor}
 import x7c1.linen.modern.init.settings.preset.AttachSourceDialog.Arguments
-import x7c1.wheat.modern.formatter.ThrowableFormatter
-import ThrowableFormatter.format
+import x7c1.linen.repository.channel.my.{ChannelToAttach, ChannelsToAttachAccessor}
 import x7c1.wheat.ancient.context.ContextualFactory
 import x7c1.wheat.ancient.resource.ViewHolderProviderFactory
 import x7c1.wheat.lore.resource.AdapterDelegatee
 import x7c1.wheat.macros.fragment.TypedFragment
 import x7c1.wheat.macros.logger.Log
 import x7c1.wheat.modern.decorator.Imports._
+import x7c1.wheat.modern.formatter.ThrowableFormatter.format
 
 import scala.collection.mutable
