+++ linen-modern/src/main/scala/x7c1/linen/modern/init/settings/preset/PresetsSelectedFragment.scala
@@ -8,8 +8,8 @@ import android.view.{LayoutInflater, Menu, MenuItem, View, ViewGroup}
 import x7c1.linen.glue.res.layout.{SettingPresetChannelRow, SettingPresetTabSelected}
 import x7c1.linen.glue.service.ServiceControl
 import x7c1.linen.glue.service.ServiceLabel.Updater
-import x7c1.linen.modern.accessor.setting.SelectedPresetChannelsAccessor
 import x7c1.linen.modern.init.updater.UpdaterMethods
+import x7c1.linen.repository.channel.preset.SelectedPresetChannelsAccessor
 import x7c1.wheat.ancient.resource.ViewHolderProviderFactory
 import x7c1.wheat.macros.fragment.TypedFragment
 import x7c1.wheat.macros.intent.ServiceCaller
