+++ linen-repository/src/main/scala/x7c1/linen/repository/source/unread/RawSourceAccessor.scala
@@ -1,7 +1,6 @@
-package x7c1.linen.modern.accessor.unread
+package x7c1.linen.repository.source.unread
 
 import x7c1.linen.database.LinenOpenHelper
-import x7c1.linen.repository.source.SourceTitle
 import x7c1.wheat.macros.logger.Log
 
 class RawSourceAccessor(helper: LinenOpenHelper){
